reorganise json folder
